requirements.txt
apache-airflow==2.0.0
xmltodict

# dev
pytest
pytest-cov
flake8
twine
